dash_app/requirements.txt
dash
dotenv
pandas
psycopg2-binary
plotly